// src/type/pull-request.type.ts
export type PRCommentType = 'GENERAL' | 'SUGGESTION' | 'QUESTION' | 'APPROVAL' | 'REQUEST_CHANGES';

export interface IPRSuggestion {
  line?: number;
  originalText?: string;
  suggestedText?: string;
}